requirements.txt
pytest
pywinauto
selenium
pytest-selenium
pyyaml
pyperclip
pyautogui
numpy
opencv-python
redis
pymongo